refactor(core): move trustzone initialization to sec layer
[no changelog]
