+++ core/site_scons/models/stm32u5_common.py
@@ -26,6 +26,7 @@ def stm32u5_common_files(env, features_wanted, defines, sources, paths):
         "embed/sec/secure_aes/inc",
         "embed/sec/tamper/inc",
         "embed/sec/time_estimate/inc",
+        "embed/sec/trustzone/inc",
         "embed/sec/unit_properties/inc",
         "embed/sys/bsp/stm32u5",
         "embed/sys/cpuid/inc",
@@ -104,6 +105,7 @@ def stm32u5_common_files(env, features_wanted, defines, sources, paths):
         "embed/sec/storage/stm32u5/storage_salt.c",
         "embed/sec/tamper/stm32u5/tamper.c",
         "embed/sec/time_estimate/stm32/time_estimate.c",
+        "embed/sec/trustzone/stm32u5/tz_init.c",
         "embed/sec/unit_properties/stm32/unit_properties.c",
         "embed/sys/cpuid/stm32/cpuid.c",
         "embed/sys/flash/stm32u5/flash.c",
