+++ core/site_scons/models/unix_common.py
@@ -16,6 +16,7 @@ def unix_common_files(env, features_wanted, defines, sources, paths):
         "embed/sec/fwutils/inc",
         "embed/sec/random_delays/inc",
         "embed/sec/time_estimate/inc",
+        "embed/sec/trustzone/inc",
         "embed/sys/bsp/inc",
         "embed/sys/inc",
         "embed/sec/board_capabilities/inc",
@@ -44,6 +45,7 @@ def unix_common_files(env, features_wanted, defines, sources, paths):
         "embed/sec/monoctr/unix/monoctr.c",
         "embed/sec/rng/rng_strong.c",
         "embed/sec/time_estimate/unix/time_estimate.c",
+        "embed/sec/trustzone/unix/tz_init.c",
         "embed/sec/unit_properties/unix/unit_properties.c",
         "embed/sys/cpuid/unix/cpuid.c",
         "embed/sys/flash/unix/flash.c",
